requirements.txt
openai>=1.40.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26
python-dotenv>=1.0.1
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
pyobjc>=10.3.1; sys_platform == 'darwin'
